Further work on Scoring
