+++ unity/Assets/Scripts/DanceSceneMain.cs
@@ -1,6 +1,8 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System.IO;
+using System.Linq;
 
 namespace PoseTeacher
 {
@@ -10,10 +12,13 @@ namespace PoseTeacher
         PoseInputGetter selfPoseInputGetter, teacherPoseInputGetter;
 
         public GameObject videoCube;
+        public GameObject scoreDisplay;
 
         public GameObject avatarContainerSelf, avatarContainerTeacher;
         List<AvatarContainer> avatarListSelf, avatarListTeacher;
 
+        private readonly string dance_file = "jsondata/salsa_m/2020_12_14-15_46_29.txt";
+        private readonly string move_file = "jsondata/move1.txt";
         private readonly string fake_file = "jsondata/2020_05_27-00_01_59.txt";
         public PoseInputSource selfPoseInputSource = PoseInputSource.KINECT;
 
@@ -23,6 +28,10 @@ namespace PoseTeacher
 
         PoseData currentSelfPose;
 
+        List<PoseData> move = new List<PoseData>();
+
+        int currentframe = 0;
+
         // Start is called before the first frame update
         void Start()
         {
@@ -31,30 +40,45 @@ namespace PoseTeacher
             avatarListSelf.Add(new AvatarContainer(avatarContainerSelf));
             avatarListTeacher.Add(new AvatarContainer(avatarContainerTeacher));
 
-            selfPoseInputGetter = new PoseInputGetter(selfPoseInputSource) { ReadDataPath = fake_file };
-            teacherPoseInputGetter = new PoseInputGetter(PoseInputSource.FILE) { ReadDataPath = fake_file };
+            selfPoseInputGetter = new PoseInputGetter(selfPoseInputSource) { ReadDataPath = dance_file };
+            teacherPoseInputGetter = new PoseInputGetter(PoseInputSource.FILE) { ReadDataPath = dance_file };
             selfPoseInputGetter.loop = true;
-            teacherPoseInputGetter.loop = true;
+            teacherPoseInputGetter.loop = false;
 
             selfPoseInputGetter.VideoCube = videoCube;
             
-            scoringUtil = new ScoringScript(avatarListSelf[0]);
+            scoringUtil = new ScoringScript(scoreDisplay);
+
+            List<string> sequence = File.ReadLines(move_file).ToList();
+            for (int i = 0; i<sequence.Count;i+=10)
+            {
+                move.Add(PoseDataUtils.JSONstring2PoseData(sequence[i]));
+            }
+            scoringUtil.StartNewGoal(GoalType.MOTION, move);
         }
 
         // Update is called once per frame
         private void Update()
         {
-            if (!paused)
-            {
-                currentSelfPose = selfPoseInputGetter.GetNextPose();
-                AnimateSelf(currentSelfPose);
-                AnimateTeacher(teacherPoseInputGetter.GetNextPose());
-            }
+            
         }
 
         private void FixedUpdate()
         {
+            
+            if (!paused)
+            {
+                currentframe += 1;
+                currentSelfPose = selfPoseInputGetter.GetNextPose();
+                AnimateSelf(currentSelfPose);
+                AnimateTeacher(teacherPoseInputGetter.GetNextPose());
 
+                if (currentframe >= 10)
+                {
+                    currentframe = 0;
+                    scoringUtil.Update(currentSelfPose);
+                }
+            }
         }
 
         private void OnApplicationQuit()
