+++ unity/Assets/Scripts/Scoring.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4802dab47b769e549bcaa4822f914b0d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
