--- unity/Assets/Scripts/ScoringScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d5c1370bf742f5448d2bab2812fc640
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
